Simplify TypeScript config for tests
After updating to latest version of ts-jest, it started showing weird warnings (see kulshekhar/ts-jest#1173), which was useless to me (since the suggested change is already done). Turns out, ts-jest was confused by the missing tsconfig.json file in the repo root. I fixed this by switching to my current convention:
- tsconfig.json used during development for richer IntelliSense;
- separate tsconfig.*.json configs for building and testing;

This change also changed the entity that is default-exported from "/src/index.ts" (previously just a function, now the whole module). Since the module is made to be callable and mimic the function itself, this is not a breaking change (only in this case; normally it would be breaking). However, I honestly don't know why I was expecting to not import the whole module, as the code in src/index.ts explicitly states to export a callable object that has properties 'retryable', 'wait', and 'default'.